Refactor tracklet
