+++ pixano/datasets/features/__init__.py
@@ -67,7 +67,6 @@ from .types import (
     ItemRef,
     NDArrayFloat,
     SchemaRef,
-    TrackRef,
     ViewRef,
     create_annotation_ref,
     create_embedding_ref,
@@ -75,7 +74,6 @@ from .types import (
     create_item_ref,
     create_ndarray_float,
     create_schema_ref,
-    create_track_ref,
     create_view_ref,
     is_annotation_ref,
     is_base_type,
@@ -84,7 +82,6 @@ from .types import (
     is_item_ref,
     is_ndarray_float,
     is_schema_ref,
-    is_track_ref,
     is_view_ref,
 )
 
@@ -120,7 +117,6 @@ __all__ = [
     "SchemaRef",
     "TrackRef",
     "Tracklet",
-    "TrackRef",
     "Video",
     "View",
     "is_annotation",
@@ -148,14 +144,12 @@ __all__ = [
     "is_item_ref",
     "is_view_ref",
     "is_entity_ref",
-    "is_track_ref",
     "is_annotation_ref",
     "is_embedding_ref",
     "create_annotation_ref",
     "create_entity_ref",
     "create_item_ref",
     "create_schema_ref",
-    "create_track_ref",
     "create_view_ref",
     "create_embedding_ref",
     "create_bbox",
