+++ pixano/datasets/features/types/__init__.py
@@ -12,21 +12,18 @@ from .schema_reference import (
     EntityRef,
     ItemRef,
     SchemaRef,
-    TrackRef,
     ViewRef,
     create_annotation_ref,
     create_embedding_ref,
     create_entity_ref,
     create_item_ref,
     create_schema_ref,
-    create_track_ref,
     create_view_ref,
     is_annotation_ref,
     is_embedding_ref,
     is_entity_ref,
     is_item_ref,
     is_schema_ref,
-    is_track_ref,
     is_view_ref,
 )
 
@@ -47,7 +44,6 @@ __all__ = [
     "create_view_ref",
     "create_annotation_ref",
     "create_schema_ref",
-    "create_track_ref",
     "create_embedding_ref",
     "is_base_type",
     "is_ndarray_float",
@@ -56,6 +52,5 @@ __all__ = [
     "is_view_ref",
     "is_annotation_ref",
     "is_schema_ref",
-    "is_track_ref",
     "is_embedding_ref",
 ]
